requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
einops==0.8.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.2.3
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.67.1
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
lightning-utilities==0.11.8
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.29.0
opencv-python==4.10.0.84
packaging==24.2
pandas
pillow==10.4.0
propcache==0.2.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scipy
six==1.16.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==1.0.11
torch==2.4.1
torchmetrics==1.5.2
torchvision==0.19.1
tqdm==4.67.0
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
